Rename grouping select box to market
